refactor models to dir and add conversion function
