+++ src/prime_rl/trainer/models/glm4_moe/__init__.py
@@ -1,0 +1,9 @@
+from prime_rl.trainer.models.glm4_moe.configuration_glm4_moe import Glm4MoeConfig
+from prime_rl.trainer.models.glm4_moe.modeling_glm4_moe import Glm4MoeForCausalLM, Glm4MoeModel, Glm4MoePreTrainedModel
+
+__all__ = [
+    "Glm4MoeConfig",
+    "Glm4MoeForCausalLM",
+    "Glm4MoeModel",
+    "Glm4MoePreTrainedModel",
+]
