+++ src/prime_rl/trainer/models/qwen3_moe/__init__.py
@@ -1,0 +1,13 @@
+from prime_rl.trainer.models.qwen3_moe.configuration_qwen3_moe import Qwen3MoeConfig
+from prime_rl.trainer.models.qwen3_moe.modeling_qwen3_moe import (
+    Qwen3MoeForCausalLM,
+    Qwen3MoeModel,
+    Qwen3MoePreTrainedModel,
+)
+
+__all__ = [
+    "Qwen3MoeConfig",
+    "Qwen3MoeForCausalLM",
+    "Qwen3MoeModel",
+    "Qwen3MoePreTrainedModel",
+]
