+++ src/prime_rl/trainer/rl/broadcast/nccl.py
@@ -13,7 +13,7 @@ from vllm.distributed.utils import StatelessProcessGroup
 from prime_rl.trainer.rl.broadcast.base import WeightBroadcast
 from prime_rl.trainer.rl.config import NCCLWeightBroadcastConfig
 from prime_rl.trainer.utils import get_world
-from prime_rl.trainer.weights import convert_tt_layer_to_hf, get_max_layer_num, has_tt_moe_layers
+from prime_rl.trainer.weights import get_max_layer_num, has_tt_moe_layers
 from prime_rl.utils.logger import get_logger
 
 
@@ -122,7 +122,7 @@ class NCCLWeightBroadcastSender:
                 state_dict[key] = value
 
             if has_tt_moe_layers(state_dict):
-                convert_tt_layer_to_hf(state_dict, layer_id)
+                model.convert_tt_layer_to_hf(state_dict, layer_id)
 
             if self.world.is_master:
                 broadcast_state_dict(state_dict, self.communicator)
